[MS-28-VideoClipPlayer][FIX]UIComponents 네임스페이싱 제거
