+++ MySeconds/Modules/MySecondsKit/Sources/CollectionViews/IntrinsicCollectionView.swift
