--- MySeconds/Modules/MySecondsKit/Sources/UIComponents.swift
@@ -1,16 +1,0 @@
-//
-//  UIComponents.swift
-//  MySeconds
-//
-//  Created by hh647 on 01/27/2025.
-//
-
-import Foundation
-
-public enum UIComponents {
-    public enum Buttons {}
-
-    public enum Labels {}
-
-    public enum Views {}
-}
